requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-rosetta==0.9.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.16.0
idna==3.2
Pillow==8.3.2
polib==1.1.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
